requirements.txt
boto3
pandas
pyarrow 
python-dotenv